Added -F option to remove bsub script. Modified -f default value to be "cufflinks". Modified derivation of -J parameter to bsub invocation. Other minor mods.

+++ bin/QuantDisplay_load_lsf.pl
@@ -1,12 +1,12 @@
-#! /usr/bin/perl
+#!/usr/bin/env perl
 #
 use strict;
 use Bio::DB::GFF;
 use Getopt::Std;
 use Cwd;
-use vars qw/ $opt_h $opt_f $opt_d $opt_u $opt_p $opt_H $opt_b $opt_r $opt_l $opt_a $opt_A $opt_R $opt_v $opt_c $opt_q $opt_s /;
+use vars qw/ $opt_h $opt_f $opt_d $opt_u $opt_p $opt_H $opt_b $opt_r $opt_l $opt_a $opt_A $opt_R $opt_v $opt_c $opt_q $opt_s $opt_F /;
 $Getopt::Std::STANDARD_HELP_VERSION = 1;
-getopts('hf:du:p:H:b:r:R:l:aAvcq:s:');
+getopts('hf:du:p:H:b:r:R:l:aAvcq:s:F');
 
 my $user = $opt_u || 'yeast';
 my $password = $opt_p || undef;
@@ -15,15 +15,18 @@ my $db = $opt_b || 'yeast_chr1';
 my $help = $opt_h || undef;
 my $debug = $opt_d || undef;
 my $verbose = $opt_v || $debug || undef;
-my $featuretype = $opt_f || 'blah';
+#my $featuretype = $opt_f || 'blah';
+my $featuretype = $opt_f || 'cufflinks';
 my $urefmol = $opt_r || 'I';
 my $usage = "QuantDisplay_load.pl -p ( -h -f -d -u -H -b -v )";
 my $refclass = $opt_R || 'chromosome';
 my $QDcount = $opt_c || undef;
 #my $queues = $opt_q;
 my $queues = 0;
+my $queue = $opt_q || 'normal';
 my $stop = $opt_s;
 my $start = $opt_l;
+my $unlink = $opt_F;
 
 my $QD_script = "/home/sgivan/projects/QuantDisplay/bin/QuantDisplay_load.pl";
 #my $QD_script = "/home/sgivan/bin/QuantDisplay_load.pl";
@@ -49,7 +52,7 @@ ie: QuantDisplay_load_lsf.pl -u brachy_cluster -b QuantDisplay01 -p password -H 
 Options
 
 -h		print this help menu
--f		feature type (default = 'blah')
+-f		feature type (default = 'cufflinks')
 -a		<feature type>
 
 			perform analysis on all features of type <feature type>
@@ -67,6 +70,8 @@ Options
 -c		for each feature, count based on QDcount attribute
 -l		with default sorting, start at this reference molecule
 -s		with default sorting, stop at this reference molecule
+-q      LSF queue to submit jobs (default = 'normal')
+-F      delete bsub script after submitting job
 
 HELP
 exit();
@@ -130,11 +135,13 @@ foreach my $refmol (@refmols) {
 	open(OUT,">$refmol" . "_QD_lsf.sh") or die "can't open '$refmol" . "QD_lsf.sh: $!";
 
     #print OUT "#\$ -o $ENV{HOME}/cluster/QDout$refmol\n#\$ -e $ENV{HOME}/cluster/QDerror$refmol\n#\$ -N $featuretype" . "$refmol\n\n";
-	print OUT "#BSUB -oo QD$refmol" . ".o\%J\n#BSUB -eo QD$refmol" . ".e\%J\n#BSUB -J $featuretype" . "$refmol\n\n";
+	print OUT "#BSUB -oo QD$refmol" . ".o\%J\n#BSUB -eo QD$refmol" . ".e\%J\n#BSUB -J " . substr($featuretype,0,4) . "$refmol\n#BSUB -q $queue\n\n";
 	
 	print OUT "$QD_script -u $user -p $password -H $dbhost -b $db -f $featuretype -r $refmol -R $refclass -c -o $gffout -X x\n\n";
 	
 	close(OUT);
+
+    #exit();
 
 	my $qsub_return;
 	unless ($debug) {
@@ -146,15 +153,17 @@ foreach my $refmol (@refmols) {
 		} else {
 #			system("qsub $refmol" . "_QD_lsf.sh");
 
-			open(QSUB, "qsub $refmol" . "_QD_lsf.sh |") or die "can't open qsub: $!";
+			#open(QSUB, "qsub $refmol" . "_QD_lsf.sh |") or die "can't open qsub: $!";
+			open(QSUB, "bsub < $refmol" . "_QD_lsf.sh |") or die "can't open qsub: $!";
 
 		}
 		
 		$qsub_return = <QSUB>;
 		close(QSUB) or warn("can't close queue: $!");
-		unlink("$refmol" . "_QD_lsf.sh");
+		unlink("$refmol" . "_QD_lsf.sh") if ($unlink);
 		
-		if (! ($cnt % $job_interval)) {
+#		if (! ($cnt % $job_interval)) {
+		if (0) {
 		
 			if ($qsub_return =~ /\sjob\s(\d+)\s/m) {
 				my $jobid = $1;
